requirements.txt
Pillow
click
flake8
numpy
scipy
tensorflow
Keras
